Take out mapState/mapDispatch to app.connect
