+++ src/components/offers-list/offers-list.connect.jsx
@@ -1,0 +1,2 @@
+
+//
